Add Tests for Bank
This will also fix the discovered bug that we operate with int and doubles

+++ src/main/java/ch/engenius/bank/Bank.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 public class Bank {
     private HashMap<Integer, Account> accounts = new HashMap<>();
 
-    public Account registerAccount(int accountNumber, int amount) {
+    public Account registerAccount(int accountNumber, double amount) {
         Account account = new Account();
         account.setMoney(amount);
         accounts.put(accountNumber, account);
